Move ServiceFinderBinder into an innate package Make internal package classes final
