+++ core-common/src/main/java/org/glassfish/jersey/innate/inject/ServiceFinderBinder.java
@@ -14,13 +14,14 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.jersey.internal;
+package org.glassfish.jersey.innate.inject;
 
 import java.util.Map;
 
 import jakarta.ws.rs.RuntimeType;
 
 import org.glassfish.jersey.innate.inject.InternalBinder;
+import org.glassfish.jersey.internal.ServiceFinder;
 import org.glassfish.jersey.internal.inject.InjectionManager;
 import org.glassfish.jersey.internal.util.PropertiesHelper;
