+++ core-common/src/main/java/org/glassfish/jersey/internal/DynamicFeatureConfigurator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, 2022 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2021, 2024 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0, which is available at
@@ -37,7 +37,7 @@ import java.util.Set;
  * in the list of configurators due to same list of {@link org.glassfish.jersey.internal.spi.AutoDiscoverable}
  * used in the {@link BootstrapBag} to register discovered features.
  */
-public class DynamicFeatureConfigurator extends AbstractFeatureConfigurator<DynamicFeature> {
+public final class DynamicFeatureConfigurator extends AbstractFeatureConfigurator<DynamicFeature> {
 
     /**
      * Create a new configurator.
