+++ core-common/src/main/java/org/glassfish/jersey/internal/Errors.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, 2018 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2024 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0, which is available at
@@ -40,7 +40,7 @@ import org.glassfish.jersey.internal.util.Producer;
  *
  * @author Michal Gajdos
  */
-public class Errors {
+public final class Errors {
 
     private static final Logger LOGGER = Logger.getLogger(Errors.class.getName());
