+++ tests/e2e-core-common/src/test/java/org/glassfish/jersey/tests/e2e/common/config/TestContract.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, 2019 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2024 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0, which is available at
@@ -16,8 +16,10 @@
 
 package org.glassfish.jersey.tests.e2e.common.config;
 
+import org.glassfish.jersey.innate.inject.ServiceFinderBinder;
+
 /**
- * Test contract for {@link org.glassfish.jersey.internal.ServiceFinderBinder} test.
+ * Test contract for {@link ServiceFinderBinder} test.
  *
  * @author Marek Potociar
  */
